requirements.txt
pillow==9.5.0
selenium==3.141.0
requests==2.28.2
opencv-python==4.8.0.76
numpy==1.24.3
PyAutoGUI==0.9.54